intial user filing system
just a preliminary filing system to users. the program reads the a csv file to a 2d vector and this vector can have users added to it using a function called write_to_users_vector and this vector is then written to a csv file, note that i am currently using absolute paths for the file being written too because you can not write to resource files in qt, this is of course just a stop gap method for testing purposes and the users are going to be stored in a map later on. and of course the writing and reading will be from the same file not as it is now